Templates: Move template preview text into image
This creates a rough preview of the template using the set text color
for the friend code and in-game name.

In addition, this refactors the card styling a bit. Most notably the
drop shadow is reduced so that it isn't overlapping with other elements.
Some rounded corners were also tightened.